refactor(api): move middleware to internal package and update health route
Add jman-api binary to release artifacts
